% Generated by build script
\name{ecb_bond_yield}
\alias{ecb_bond_yield}
\title{Euro area government bond yields}
\usage{
ecb_bond_yield()
}
\arguments{
\item{maturity}{Maturity: "1Y", "2Y", "5Y", "10Y", "30Y"}
\item{last_n}{Last N observations (default 60)}
}
\value{tibble: date, value, key, title}
\description{Euro area government bond yields}
