% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecb.R
\name{ecb_exchange_rate}
\alias{ecb_exchange_rate}
\title{ECB reference exchange rates}
\usage{
ecb_exchange_rate(currency = "USD", start = NULL, end = NULL, last_n = NULL)
}
\arguments{
\item{currency}{3-letter currency code (e.g. "USD", "GBP", "JPY", "CHF").
Use "." for all currencies.}

\item{start}{Start date}

\item{end}{End date}

\item{last_n}{Last N observations}
}
\value{
tibble: date, value, currency, title
}
\description{
Daily EUR exchange rates against major currencies.
}
