% Generated by build script
\name{ecb_exchange_rates}
\alias{ecb_exchange_rates}
\title{Exchange rates for multiple currencies}
\usage{
ecb_exchange_rates()
}
\arguments{
\item{currencies}{Character vector of currency codes}
\item{last_n}{Last N observations per currency (default 30)}
}
\value{tibble: date, value, currency, title}
\description{Exchange rates for multiple currencies}
