% Generated by build script
\name{ecb_inflation}
\alias{ecb_inflation}
\title{Euro area HICP inflation}
\usage{
ecb_inflation()
}
\arguments{
\item{component}{"000000" (all items, default), "010000" (food),}
\item{last_n}{Last N observations (default 60)}
}
\value{tibble: date, value, key, title}
\description{Euro area HICP inflation}
