% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecb.R
\name{ecb_key_rates}
\alias{ecb_key_rates}
\title{ECB key interest rates (main refinancing operations, deposit facility, etc.)}
\usage{
ecb_key_rates(last_n = 20)
}
\arguments{
\item{last_n}{Last N observations (default 20)}
}
\value{
tibble: date, value, key, title
}
\description{
ECB key interest rates (main refinancing operations, deposit facility, etc.)
}
