% Generated by build script
\name{ecb_exchange_rate}
\alias{ecb_exchange_rate}
\title{ECB reference exchange rates}
\usage{
ecb_exchange_rate()
}
\arguments{
\item{currency}{3-letter currency code (e.g. "USD", "GBP", "JPY", "CHF").}
\item{start}{Start date}
\item{end}{End date}
\item{last_n}{Last N observations}
}
\value{tibble: date, value, currency, title}
\description{ECB reference exchange rates}
